Adding test coverage.
