+++ test/functional_test/test_models/test_kafka_consumer.py
@@ -93,7 +93,3 @@ class TestKafkaConsumerFunctional(unittest.TestCase):
 
         mock_logging.info.assert_called()
         consumer_instance.close.assert_called_once()
-
-
-if __name__ == "__main__":
-    unittest.main()
